refactor(location-manager): decompose attractions form
