+++ apps/location-manager/packages/client/src/features/location-create/components/attractions-form/attractionsForm.types.ts
@@ -1,0 +1,6 @@
+export type AttractionsFormSection =
+  | "step1"
+  | "entities"
+  | "profile"
+  | "tours"
+  | "visitContact";
